tetragon: Move argument processing code to args.go
Moving argument processing code to args.go structure the code better,
because it will be used also from uprobe code. Moving it together with
two other helper functions.
